backend/requirements.txt
Django==2.1
psycopg2
django-extensions==2.1.0
redis==2.10.3
django-redis==4.9.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
elasticsearch-dsl==6.1.0
django-elasticsearch-dsl==0.5.0
djoser==1.2.1
django-cors-headers==2.4.0